Form validation, confirm before delete event

+++ src/lib/postSave.js
@@ -3,49 +3,16 @@ const plugins = require.main.require('./src/plugins');
 
 import validator from 'validator';
 import Promise from 'bluebird';
-import { default as parse } from './parse';
+import parse from './parse';
 import { canPostEvent } from './privileges';
-import { deleteEvent, saveEvent } from './event';
+import { deleteEvent, saveEvent, eventExists, getEvent } from './event';
+import validateEvent from './validateEvent';
+import { notify } from './reminders';
 
 const log = (...args) => console.log(...args);
 const p = Promise.promisify;
 
 const fireHook = p(plugins.fireHook);
-
-const isArrayOf = (arr, type) => {
-  const isType = typeof type === 'function' ? type : it => typeof it === type;
-  if (!Array.isArray(arr)) {
-    return false;
-  }
-  for (const x of arr) {
-    if (!isType(x)) {
-      return false;
-    }
-  }
-  return true;
-};
-
-const validateEvent = event => {
-  const l = (bool, message) => {
-    if (!bool) {
-      log('[plugin-calendar] Event validation failed at ', message);
-    }
-    return bool;
-  };
-
-  if (
-    l(typeof event.name === 'string', 'name') &&
-    l(typeof event.allday === 'boolean', 'allday') &&
-    l(new Date(event.startDate).getTime() === event.startDate, 'startDate') &&
-    l(new Date(event.endDate).getTime() === event.endDate, 'endDate') &&
-    l(isArrayOf(event.reminders, 'number'), 'reminders') &&
-    l(typeof event.location === 'string', 'location') &&
-    l(typeof event.description === 'string', 'description')
-  ) {
-    return event;
-  }
-  return null;
-};
 
 const regex = new RegExp(
   '(\\[\\s?event\\s?\\][\\w\\W]*\\[\\s?\\/\\s?event\\s?\\])|' +
@@ -57,28 +24,53 @@ const postSave = async postData => {
 
   // delete event if no longer in post
   if (!postData.content.match(regex)) {
-    await deleteEvent(postData.pid);
-    log(`[plugin-calendar] Event (pid:${postData.pid}) saved`);
+    const existed = await eventExists(postData.pid);
+    if (existed) {
+      await notify({
+        event: await getEvent(postData.pid),
+        message: '[[calendar:event_deleted]]',
+      });
+
+      await deleteEvent(postData.pid);
+      log(`[plugin-calendar] Event (pid:${postData.pid}) deleted`);
+    }
 
     return postData;
   }
 
-  event = validateEvent(event);
-  if (!event || !(await canPostEvent(postData.pid, postData.uid))) {
-    return {
-      ...postData,
-      content: postData.content.replace(
-        /\[\s?(\/?)\s?event\s?\]/g,
-        '[$1event-invalid]'
-      ),
-    };
+  const invalid = () => ({
+    ...postData,
+    content: postData.content.replace(
+      /\[\s?(\/?)\s?event\s?\]/g,
+      '[$1event-invalid]'
+    ),
+  });
+
+  if (!event) {
+    return invalid();
+  }
+
+  const [failed, failures] = validateEvent(event);
+  if (failed) {
+    const obj = failures.reduce((val, failure) => ({
+      ...val,
+      [failure]: event[failure],
+    }), {});
+    log(`[plugin-calendar] Event (pid:${postData.pid}) validation failed: `, obj);
+    return invalid();
+  }
+
+  const can = await canPostEvent(postData.pid, postData.uid);
+  if (!can) {
+    return invalid();
   }
 
   event.name = validator.escape(event.name);
   event.pid = postData.pid;
   event.tid = postData.tid;
+  event.cid = postData.cid;
   event.uid = postData.uid;
-  event = (await fireHook('filter:plugin-calendar:event.post', event));
+  event = await fireHook('filter:plugin-calendar:event.post', event);
 
   await saveEvent(event);
   log(`[plugin-calendar] Event (pid:${event.pid}) saved`);
